solve linear equations with one/ two variables - solution must be generalized for every variable notation
