+++ src/fixtures/latex-equal/symbolic/equation-with-no-variables.ts
@@ -3,14 +3,9 @@ export default {
   tests: [
     { target: "2+2=4", ne: ["1", "2=2"] },
     {
-      // only: true,
       target: "4=4",
-      // eq: ["4=4", "2+2=4"],
-      ne: [
-        // "1",
-        // "3=3", "9=8", "a=b", "0=0",
-        "\\true=\\true",
-      ],
+      eq: ["4=4", "2+2=4"],
+      ne: ["1", "3=3", "9=8", "a=b", "0=0", "\\true=\\true"],
     },
     {
       target: "4=4=2+2",
