+++ src/fixtures/latex-equal/symbolic/math-series.ts
@@ -1,6 +1,5 @@
 export default {
   mode: "symbolic",
-  //only: true,
   tests: [
     {
       target: "a_n=3a_{n-1}",
